+++ src/mock/MockResolver.sol
@@ -8,14 +8,14 @@ import { OwnableUpgradeable } from "@openzeppelin/contracts-upgradeable/access/O
 contract MockResolverAdmin is OwnableUpgradeable {
     using SafeERC20 for IERC20;
 
-    mapping(uint256 schemaId => uint256 ethFees) public schemaAttestETHFees;
-    mapping(uint256 schemaId => mapping(IERC20 tokenAddress => uint256 tokenFees)) public schemaAttestTokenFees;
-    mapping(uint256 attestationId => uint256 ethFees) public attestationETHFees;
-    mapping(uint256 attestationId => mapping(IERC20 tokenAddress => uint256 tokenFees)) public attestationTokenFees;
+    mapping(uint64 schemaId => uint256 ethFees) public schemaAttestETHFees;
+    mapping(uint64 schemaId => mapping(IERC20 tokenAddress => uint256 tokenFees)) public schemaAttestTokenFees;
+    mapping(uint64 attestationId => uint256 ethFees) public attestationETHFees;
+    mapping(uint64 attestationId => mapping(IERC20 tokenAddress => uint256 tokenFees)) public attestationTokenFees;
     mapping(IERC20 tokenAddress => bool approved) public approvedTokens;
 
-    event ETHFeesReceived(uint256 attestationId, uint256 amount);
-    event TokenFeesReceived(uint256 attestationId, IERC20 token, uint256 amount);
+    event ETHFeesReceived(uint64 attestationId, uint256 amount);
+    event TokenFeesReceived(uint64 attestationId, IERC20 token, uint256 amount);
 
     error MismatchETHFee();
     error InsufficientETHFee();
@@ -26,11 +26,11 @@ contract MockResolverAdmin is OwnableUpgradeable {
         __Ownable_init(_msgSender());
     }
 
-    function setSchemaAttestETHFees(uint256 schemaId, uint256 fees) external onlyOwner {
+    function setSchemaAttestETHFees(uint64 schemaId, uint256 fees) external onlyOwner {
         schemaAttestETHFees[schemaId] = fees;
     }
 
-    function setSchemaAttestTokenFees(uint256 schemaId, IERC20 token, uint256 fees) external onlyOwner {
+    function setSchemaAttestTokenFees(uint64 schemaId, IERC20 token, uint256 fees) external onlyOwner {
         schemaAttestTokenFees[schemaId][token] = fees;
     }
 
@@ -38,7 +38,7 @@ contract MockResolverAdmin is OwnableUpgradeable {
         approvedTokens[token] = approved;
     }
 
-    function _receiveEther(address attester, uint256 schemaId, uint256 attestationId) internal {
+    function _receiveEther(address attester, uint64 schemaId, uint64 attestationId) internal {
         uint256 fees =
             schemaAttestETHFees[schemaId] == 0 ? attestationETHFees[attestationId] : schemaAttestETHFees[schemaId];
         if (msg.value != fees) revert InsufficientETHFee();
@@ -48,8 +48,8 @@ contract MockResolverAdmin is OwnableUpgradeable {
 
     function _receiveTokens(
         address attester,
-        uint256 schemaId,
-        uint256 attestationId,
+        uint64 schemaId,
+        uint64 attestationId,
         IERC20 resolverFeeERC20Token,
         uint256 resolverFeeERC20Amount
     )
@@ -68,8 +68,8 @@ contract MockResolverAdmin is OwnableUpgradeable {
 contract MockResolver is ISPHook, MockResolverAdmin {
     function didReceiveAttestation(
         address attester,
-        uint256 schemaId,
-        uint256 attestationId,
+        uint64 schemaId,
+        uint64 attestationId,
         bytes calldata
     )
         external
@@ -80,8 +80,8 @@ contract MockResolver is ISPHook, MockResolverAdmin {
 
     function didReceiveAttestation(
         address attester,
-        uint256 schemaId,
-        uint256 attestationId,
+        uint64 schemaId,
+        uint64 attestationId,
         IERC20 resolverFeeERC20Token,
         uint256 resolverFeeERC20Amount,
         bytes calldata
@@ -94,8 +94,8 @@ contract MockResolver is ISPHook, MockResolverAdmin {
 
     function didReceiveRevocation(
         address attester,
-        uint256 schemaId,
-        uint256 attestationId,
+        uint64 schemaId,
+        uint64 attestationId,
         bytes calldata
     )
         external
@@ -107,8 +107,8 @@ contract MockResolver is ISPHook, MockResolverAdmin {
 
     function didReceiveRevocation(
         address attester,
-        uint256 schemaId,
-        uint256 attestationId,
+        uint64 schemaId,
+        uint64 attestationId,
         IERC20 resolverFeeERC20Token,
         uint256 resolverFeeERC20Amount,
         bytes calldata
